builder/requirements.txt
watchdog
requests
faiss-cpu
numpy
